requirements.txt
pymodi
pandas
pydub
simpleaudio
uvicorn
fastapi
jinja2

